+++ generic/tclCmdAH.c
@@ -60,7 +60,7 @@ static const char *	GetTypeFromMode(int mode);
 static int		StoreStatData(Tcl_Interp *interp, Tcl_Obj *varName,
 			    Tcl_StatBuf *statPtr);
 static int	EachloopCmd(Tcl_Interp *interp, int collect,
-			    int objc, Tcl_Obj *const objv[]);
+			    size_t objc, Tcl_Obj *const objv[]);
 static Tcl_NRPostProc	CatchObjCmdCallback;
 static Tcl_NRPostProc	ExprCallback;
 static Tcl_NRPostProc	ForSetupCallback;
@@ -121,7 +121,7 @@ int
 Tcl_BreakObjCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,		/* Current interpreter. */
-    int objc,			/* Number of arguments. */
+    size_t objc,			/* Number of arguments. */
     Tcl_Obj *const objv[])	/* Argument objects. */
 {
     if (objc != 1) {
@@ -152,7 +152,7 @@ int
 Tcl_CatchObjCmd(
     ClientData clientData,
     Tcl_Interp *interp,		/* Current interpreter. */
-    int objc,			/* Number of arguments. */
+    size_t objc,			/* Number of arguments. */
     Tcl_Obj *const objv[])	/* Argument objects. */
 {
     return Tcl_NRCallObjProc2(interp, TclNRCatchObjCmd, clientData, objc, objv);
@@ -162,7 +162,7 @@ int
 TclNRCatchObjCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,		/* Current interpreter. */
-    int objc,			/* Number of arguments. */
+    size_t objc,			/* Number of arguments. */
     Tcl_Obj *const objv[])	/* Argument objects. */
 {
     Tcl_Obj *varNamePtr = NULL;
@@ -257,7 +257,7 @@ int
 Tcl_CdObjCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,		/* Current interpreter. */
-    int objc,			/* Number of arguments. */
+    size_t objc,			/* Number of arguments. */
     Tcl_Obj *const objv[])	/* Argument objects. */
 {
     Tcl_Obj *dir;
@@ -312,7 +312,7 @@ int
 Tcl_ConcatObjCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,		/* Current interpreter. */
-    int objc,			/* Number of arguments. */
+    size_t objc,			/* Number of arguments. */
     Tcl_Obj *const objv[])	/* Argument objects. */
 {
     if (objc >= 2) {
@@ -346,7 +346,7 @@ int
 Tcl_ContinueObjCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,		/* Current interpreter. */
-    int objc,			/* Number of arguments. */
+    size_t objc,			/* Number of arguments. */
     Tcl_Obj *const objv[])	/* Argument objects. */
 {
     if (objc != 1) {
@@ -406,7 +406,7 @@ int
 EncodingConvertfromObjCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,		/* Current interpreter. */
-    int objc,			/* Number of arguments. */
+    size_t objc,			/* Number of arguments. */
     Tcl_Obj *const objv[])	/* Argument objects. */
 {
     Tcl_Obj *data;		/* Byte array to convert */
@@ -500,7 +500,7 @@ int
 EncodingConverttoObjCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,		/* Current interpreter. */
-    int objc,			/* Number of arguments. */
+    size_t objc,			/* Number of arguments. */
     Tcl_Obj *const objv[])	/* Argument objects. */
 {
     Tcl_Obj *data;		/* String to convert */
@@ -594,7 +594,7 @@ int
 EncodingDirsObjCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,		/* Current interpreter. */
-    int objc,			/* Number of arguments. */
+    size_t objc,			/* Number of arguments. */
     Tcl_Obj *const objv[])	/* Argument objects. */
 {
     Tcl_Obj *dirListObj;
@@ -638,7 +638,7 @@ int
 EncodingNamesObjCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp* interp,	    /* Tcl interpreter */
-    int objc,		    /* Number of command line args */
+    size_t objc,		    /* Number of command line args */
     Tcl_Obj* const objv[])  /* Vector of command line args */
 {
     if (objc > 1) {
@@ -669,7 +669,7 @@ int
 EncodingSystemObjCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp* interp,     /* Tcl interpreter */
-    int objc,		    /* Number of command line args */
+    size_t objc,		    /* Number of command line args */
     Tcl_Obj* const objv[])  /* Vector of command line args */
 {
     if (objc > 2) {
@@ -706,7 +706,7 @@ int
 Tcl_ErrorObjCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,		/* Current interpreter. */
-    int objc,			/* Number of arguments. */
+    size_t objc,			/* Number of arguments. */
     Tcl_Obj *const objv[])	/* Argument objects. */
 {
     Tcl_Obj *options, *optName;
@@ -768,7 +768,7 @@ int
 Tcl_EvalObjCmd(
     ClientData clientData,
     Tcl_Interp *interp,		/* Current interpreter. */
-    int objc,			/* Number of arguments. */
+    size_t objc,			/* Number of arguments. */
     Tcl_Obj *const objv[])	/* Argument objects. */
 {
     return Tcl_NRCallObjProc2(interp, TclNREvalObjCmd, clientData, objc, objv);
@@ -778,7 +778,7 @@ int
 TclNREvalObjCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,		/* Current interpreter. */
-    int objc,			/* Number of arguments. */
+    size_t objc,			/* Number of arguments. */
     Tcl_Obj *const objv[])	/* Argument objects. */
 {
     Tcl_Obj *objPtr;
@@ -837,7 +837,7 @@ int
 Tcl_ExitObjCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,		/* Current interpreter. */
-    int objc,			/* Number of arguments. */
+    size_t objc,			/* Number of arguments. */
     Tcl_Obj *const objv[])	/* Argument objects. */
 {
     Tcl_WideInt value;
@@ -884,7 +884,7 @@ int
 Tcl_ExprObjCmd(
     ClientData clientData,
     Tcl_Interp *interp,		/* Current interpreter. */
-    int objc,			/* Number of arguments. */
+    size_t objc,			/* Number of arguments. */
     Tcl_Obj *const objv[])	/* Argument objects. */
 {
     return Tcl_NRCallObjProc2(interp, TclNRExprObjCmd, clientData, objc, objv);
@@ -894,7 +894,7 @@ int
 TclNRExprObjCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,		/* Current interpreter. */
-    int objc,			/* Number of arguments. */
+    size_t objc,			/* Number of arguments. */
     Tcl_Obj *const objv[])	/* Argument objects. */
 {
     Tcl_Obj *resultPtr, *objPtr;
@@ -1030,7 +1030,7 @@ static int
 FileAttrAccessTimeCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
     Tcl_StatBuf buf;
@@ -1112,7 +1112,7 @@ static int
 FileAttrModifyTimeCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
     Tcl_StatBuf buf;
@@ -1191,7 +1191,7 @@ static int
 FileAttrLinkStatCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
     Tcl_StatBuf buf;
@@ -1227,7 +1227,7 @@ static int
 FileAttrStatCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
     Tcl_StatBuf buf;
@@ -1263,7 +1263,7 @@ static int
 FileAttrTypeCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
     Tcl_StatBuf buf;
@@ -1301,7 +1301,7 @@ static int
 FileAttrSizeCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
     Tcl_StatBuf buf;
@@ -1338,7 +1338,7 @@ static int
 FileAttrIsDirectoryCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
     Tcl_StatBuf buf;
@@ -1376,7 +1376,7 @@ static int
 FileAttrIsExecutableCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
     if (objc != 2) {
@@ -1407,7 +1407,7 @@ static int
 FileAttrIsExistingCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
     if (objc != 2) {
@@ -1438,7 +1438,7 @@ static int
 FileAttrIsFileCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
     Tcl_StatBuf buf;
@@ -1476,7 +1476,7 @@ static int
 FileAttrIsOwnedCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
 #ifdef __CYGWIN__
@@ -1523,7 +1523,7 @@ static int
 FileAttrIsReadableCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
     if (objc != 2) {
@@ -1554,7 +1554,7 @@ static int
 FileAttrIsWritableCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
     if (objc != 2) {
@@ -1585,7 +1585,7 @@ static int
 PathDirNameCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
     Tcl_Obj *dirPtr;
@@ -1624,7 +1624,7 @@ static int
 PathExtensionCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
     Tcl_Obj *dirPtr;
@@ -1663,7 +1663,7 @@ static int
 PathRootNameCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
     Tcl_Obj *dirPtr;
@@ -1702,7 +1702,7 @@ static int
 PathTailCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
     Tcl_Obj *dirPtr;
@@ -1741,7 +1741,7 @@ static int
 PathFilesystemCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
     Tcl_Obj *fsInfo;
@@ -1782,7 +1782,7 @@ static int
 PathJoinCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
     if (objc < 2) {
@@ -1814,7 +1814,7 @@ static int
 PathNativeNameCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
     Tcl_DString ds;
@@ -1851,7 +1851,7 @@ static int
 PathNormalizeCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
     Tcl_Obj *fileName;
@@ -1889,7 +1889,7 @@ static int
 PathSplitCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
     Tcl_Obj *res;
@@ -1932,7 +1932,7 @@ static int
 PathTypeCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
     Tcl_Obj *typeName;
@@ -1980,7 +1980,7 @@ static int
 FilesystemSeparatorCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
     if (objc < 1 || objc > 2) {
@@ -2035,7 +2035,7 @@ static int
 FilesystemVolumesCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
     if (objc != 1) {
@@ -2295,7 +2295,7 @@ int
 Tcl_ForObjCmd(
     ClientData clientData,
     Tcl_Interp *interp,		/* Current interpreter. */
-    int objc,			/* Number of arguments. */
+    size_t objc,			/* Number of arguments. */
     Tcl_Obj *const objv[])	/* Argument objects. */
 {
     return Tcl_NRCallObjProc2(interp, TclNRForObjCmd, clientData, objc, objv);
@@ -2305,7 +2305,7 @@ int
 TclNRForObjCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,		/* Current interpreter. */
-    int objc,			/* Number of arguments. */
+    size_t objc,			/* Number of arguments. */
     Tcl_Obj *const objv[])	/* Argument objects. */
 {
     Interp *iPtr = (Interp *) interp;
@@ -2489,7 +2489,7 @@ int
 Tcl_ForeachObjCmd(
     ClientData clientData,
     Tcl_Interp *interp,		/* Current interpreter. */
-    int objc,			/* Number of arguments. */
+    size_t objc,			/* Number of arguments. */
     Tcl_Obj *const objv[])	/* Argument objects. */
 {
     return Tcl_NRCallObjProc2(interp, TclNRForeachCmd, clientData, objc, objv);
@@ -2499,7 +2499,7 @@ int
 TclNRForeachCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
     return EachloopCmd(interp, TCL_EACH_KEEP_NONE, objc, objv);
@@ -2509,7 +2509,7 @@ int
 Tcl_LmapObjCmd(
     ClientData clientData,
     Tcl_Interp *interp,		/* Current interpreter. */
-    int objc,			/* Number of arguments. */
+    size_t objc,			/* Number of arguments. */
     Tcl_Obj *const objv[])	/* Argument objects. */
 {
     return Tcl_NRCallObjProc2(interp, TclNRLmapCmd, clientData, objc, objv);
@@ -2519,7 +2519,7 @@ int
 TclNRLmapCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
     return EachloopCmd(interp, TCL_EACH_COLLECT, objc, objv);
@@ -2531,7 +2531,7 @@ EachloopCmd(
 				 * evaluation. */
     int collect,		/* Select collecting or accumulating mode
 				 * (TCL_EACH_*) */
-    int objc,			/* The arguments being passed in... */
+    size_t objc,			/* The arguments being passed in... */
     Tcl_Obj *const objv[])
 {
     int numLists = (objc-2) / 2;
@@ -2807,7 +2807,7 @@ int
 Tcl_FormatObjCmd(
     TCL_UNUSED(ClientData),
     Tcl_Interp *interp,		/* Current interpreter. */
-    int objc,			/* Number of arguments. */
+    size_t objc,			/* Number of arguments. */
     Tcl_Obj *const objv[])	/* Argument objects. */
 {
     Tcl_Obj *resultPtr;		/* Where result is stored finally. */
