+++ generic/tclObj.c
@@ -4330,7 +4330,7 @@ int
 Tcl_RepresentationCmd(
     TCL_UNUSED(void *),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const objv[])
 {
     Tcl_Obj *descObj;
